requirements.txt
pyTelegramBotAPI==4.10.0
requests==2.28.2
denv==2.0.1
python-dotenv==0.21.1
peewee==3.15.4